requirements.txt
progressbar2
gdown
gitpython
git+https://github.com/cheind/py-thin-plate-spline
hickle
tensorboard
numpy
gradio
opencv-python
matplotlib
pyyaml
av
openmim
tqdm
psutil
open3d
pycocotools
Pillow
scikit-image
zip
ffmpeg
h5py
